Update 'empty' class description in cache model
The explanation for the 'empty' class in the cache model has been expanded to provide more context. The class is utilized to represent a lack of data in either input or output, differentiating it from a 'None' value. Furthermore, instances of the 'empty' class are used as default values in certain operations, like cache lookups. This makes handling "no data" cases more explicit and less prone to error.